+++ src/main/java/com/authlete/common/dto/IntrospectionResponse.java
@@ -218,7 +218,7 @@ import com.authlete.common.util.Utils;
  */
 public class IntrospectionResponse extends ApiResponse
 {
-    private static final long serialVersionUID = 8L;
+    private static final long serialVersionUID = 9L;
 
 
     /**
@@ -362,7 +362,7 @@ public class IntrospectionResponse extends ApiResponse
      * The content of the {@code "authorization_details"} request parameter which was
      * included in the request that obtained the token.
      */
-    private AuthzDetailsElement[] authorizationDetails;
+    private AuthzDetails authorizationDetails;
 
 
     /**
@@ -739,7 +739,7 @@ public class IntrospectionResponse extends ApiResponse
      *
      * @since 2.56
      */
-    public AuthzDetailsElement[] getAuthorizationDetails()
+    public AuthzDetails getAuthorizationDetails()
     {
         return authorizationDetails;
     }
@@ -750,13 +750,13 @@ public class IntrospectionResponse extends ApiResponse
      * {@code "authorization_details"} request parameter which is defined in
      * <i>"OAuth 2.0 Rich Authorization Requests"</i>.
      *
-     * @param elements
+     * @param details
      *         Authorization details.
      *
      * @since 2.56
      */
-    public void setAuthorizationDetails(AuthzDetailsElement[] elements)
+    public void setAuthorizationDetails(AuthzDetails details)
     {
-        this.authorizationDetails = elements;
+        this.authorizationDetails = details;
     }
 }